+++ src/main/java/dataStructure/containers/MultipleImageContainerSingleFile.java
@@ -76,7 +76,7 @@ public class MultipleImageContainerSingleFile extends MultipleImageContainer {
         ImageIOCoordinates ioCoordinates = getImageIOCoordinates(timePoint, channel);
         if (bounds!=null) ioCoordinates.setBounds(bounds);
         Image image = getReader().openImage(ioCoordinates);
-        if (scaleXY!=0 && scaleZ!=0) image.setCalibration(scaleXY, scaleZ);
+        if (scaleXY!=0 && scaleZ!=0) image.setCalibration((float)scaleXY, (float)scaleZ);
         else {
             scaleXY = image.getScaleXY();
             scaleZ = image.getScaleZ();
@@ -91,7 +91,7 @@ public class MultipleImageContainerSingleFile extends MultipleImageContainer {
         ImageIOCoordinates ioCoords = ioCoordinates.duplicate();
         ioCoords.setBounds(bounds);
         Image image = getReader().openImage(ioCoordinates);
-        if (scaleXY!=0 && scaleZ!=0) image.setCalibration(scaleXY, scaleZ);
+        if (scaleXY!=0 && scaleZ!=0) image.setCalibration((float)scaleXY, (float)scaleZ);
         else {
             scaleXY = image.getScaleXY();
             scaleZ = image.getScaleZ();
